+++ app/Http/Controllers/TestController.php
@@ -2,17 +2,65 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 
-use App\Models\User;
+use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
 
 class TestController extends Controller
 {
-    function test(){
-        $User = User::find(2);
+    /**
+     * Display a listing of the resource.
+     */
+    public function index()
+    {
+        return view("index",["name"=> "arturo", "age" => 22, 'html' => '<h1>Titulo</h1>', 'array' => [1,2,3,4]]);
+    }
 
-        echo"Hola mundo";
-        
-        return view('welcome', ['user'=>$User]);
+    /**
+     * Show the form for creating a new resource.
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     */
+    public function show(string $id)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     */
+    public function edit(string $id)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     */
+    public function update(Request $request, string $id)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     */
+    public function destroy(string $id)
+    {
+        //
     }
 }
