--- app/Http/Controllers/TextController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class TextController extends Controller
-{
-    //
-}
