+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\TestController;
 use Illuminate\Support\Facades\Route;
 //use App\Http\Controllers\TestController;
 /*
@@ -13,20 +14,5 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get("/", [App\Http\Controllers\TestController::class, 'test']);
+Route::get("/", [TestController::class,"index"]);
 
-Route::get('/custom', function () {
-    $msj = "mensaje desde el servidor c:";
-
-    $data = ['msj' => $msj,'edad'=> 15];
-    return view('custom', $data);
-
-});
-
-Route::get('/bienvenida', function () {
-    return view('welcome');
-});
-
-Route::get('/escribeme', function () {
-    return "contactame";
-})->name("contacto");
